+++ lib/ash/dsl/section.ex
@@ -25,9 +25,11 @@ defmodule Ash.Dsl.Section do
     imports: [],
     schema: [],
     describe: "",
+    snippet: "",
     examples: [],
     modules: [],
     entities: [],
-    sections: []
+    sections: [],
+    docs: ""
   ]
 end