+++ lib/ash/resource/dsl.ex
@@ -281,7 +281,7 @@ defmodule Ash.Resource.Dsl do
       """
     ],
     imports: [
-      Module.concat(["Ash", Filter.TemplateHelpers])
+      Ash.Filter.TemplateHelpers
     ],
     entities: [
       @has_one,
@@ -291,7 +291,7 @@ defmodule Ash.Resource.Dsl do
     ]
   }
 
-  @change %Ash.Dsl.Entity{
+  @action_change %Ash.Dsl.Entity{
     name: :change,
     describe: """
     A change to be applied to the changeset after it is generated. They are run in order, from top to bottom.
@@ -314,7 +314,7 @@ defmodule Ash.Resource.Dsl do
     ],
     target: Ash.Resource.Change,
     transform: {Ash.Resource.Change, :transform, []},
-    schema: Ash.Resource.Change.schema(),
+    schema: Ash.Resource.Change.action_schema(),
     args: [:change]
   }
 
@@ -352,6 +352,33 @@ defmodule Ash.Resource.Dsl do
     args: [:name, :type],
     modules: [:type],
     schema: Ash.Resource.Actions.Metadata.schema()
+  }
+
+  @change %Ash.Dsl.Entity{
+    name: :change,
+    describe: """
+    A change to be applied to the changeset after it is generated. They are run in order, from top to bottom.
+
+    To implement your own, see `Ash.Resource.Change`.
+    To use it, you can simply refer to the module and its options, like so:
+
+    `change {MyChange, foo: 1}`
+
+    But for readability, you may want to define a function elsewhere and import it,
+    so you can say something like:
+
+    `change my_change(1)`
+
+    For destroys, `changes` are not applied unless `soft?` is set to true.
+    """,
+    examples: [
+      "change relate_actor(:reporter)",
+      "change {MyCustomChange, :foo}"
+    ],
+    target: Ash.Resource.Change,
+    transform: {Ash.Resource.Change, :transform, []},
+    schema: Ash.Resource.Change.schema(),
+    args: [:change]
   }
 
   @validate %Ash.Dsl.Entity{
@@ -399,7 +426,7 @@ defmodule Ash.Resource.Dsl do
     schema: Ash.Resource.Actions.Create.opt_schema(),
     entities: [
       changes: [
-        @change,
+        @action_change,
         @action_validate
       ],
       arguments: [
@@ -467,7 +494,7 @@ defmodule Ash.Resource.Dsl do
     ],
     entities: [
       changes: [
-        @change,
+        @action_change,
         @action_validate
       ],
       metadata: [
@@ -496,7 +523,7 @@ defmodule Ash.Resource.Dsl do
     ],
     entities: [
       changes: [
-        @change,
+        @action_change,
         @action_validate
       ],
       metadata: [
@@ -534,14 +561,12 @@ defmodule Ash.Resource.Dsl do
     arguments. This does, however, allow you to customize exactly how related entities are read/
     created.
     """,
-    imports:
-      [
-        Resource.Change.Builtins,
-        Resource.Preparation.Builtins,
-        Resource.Validation.Builtins,
-        Filter.TemplateHelpers
-      ]
-      |> Enum.map(&Module.concat(["Ash", &1])),
+    imports: [
+      Ash.Resource.Change.Builtins,
+      Ash.Resource.Preparation.Builtins,
+      Ash.Resource.Validation.Builtins,
+      Ash.Filter.TemplateHelpers
+    ],
     schema: [
       defaults: [
         type: {:list, {:in, [:create, :read, :update, :destroy]}},
@@ -644,7 +669,7 @@ defmodule Ash.Resource.Dsl do
       end
       """
     ],
-    imports: [Module.concat(["Ash", Filter, TemplateHelpers])],
+    imports: [Ash.Filter.TemplateHelpers],
     schema: [
       description: [
         type: :string,
@@ -725,9 +750,7 @@ defmodule Ash.Resource.Dsl do
     describe: """
     Declare validations prior to performing actions against the resource
     """,
-    imports: [
-      Module.concat(["Ash", Resource, Validation, Builtins])
-    ],
+    imports: [Ash.Resource.Validation.Builtins],
     examples: [
       """
       validations do
@@ -738,6 +761,25 @@ defmodule Ash.Resource.Dsl do
     ],
     entities: [
       @validate
+    ]
+  }
+
+  @changes %Ash.Dsl.Section{
+    name: :changes,
+    describe: """
+    Declare changes that occur on create/update/destroy actions against the resource
+    """,
+    imports: [Ash.Resource.Validation.Builtins],
+    examples: [
+      """
+      changes do
+        change {Mod, [foo: :bar]}
+        change set_context(%{some: :context})
+      end
+      """
+    ],
+    entities: [
+      @change
     ]
   }
 
@@ -841,9 +883,7 @@ defmodule Ash.Resource.Dsl do
       end
       """
     ],
-    imports: [
-      Module.concat(["Ash", Filter.TemplateHelpers])
-    ],
+    imports: [Ash.Filter.TemplateHelpers],
     entities: [
       @count,
       @first,
@@ -923,10 +963,7 @@ defmodule Ash.Resource.Dsl do
       end
       """
     ],
-    imports: [
-      Module.concat(["Ash", Resource, Calculation, Builtins]),
-      Module.concat(["Ash", Filter.TemplateHelpers])
-    ],
+    imports: [Ash.Resource.Calculation.Builtins, Ash.Filter.TemplateHelpers],
     entities: [
       @calculation
     ]
@@ -993,6 +1030,7 @@ defmodule Ash.Resource.Dsl do
     @relationships,
     @actions,
     @resource,
+    @changes,
     @validations,
     @aggregates,
     @calculations,