+++ lib/ash/resource/preparation/preparation.ex
@@ -19,7 +19,7 @@ defmodule Ash.Resource.Preparation do
   def schema do
     [
       preparation: [
-        type: {:custom, __MODULE__, :preparation, []},
+        type: {:ash_behaviour, Ash.Resource.Preparation, Ash.Resource.Preparation.Builtins},
         doc: """
         The module and options for a preparation.
         """,