+++ lib/ash/resource/transformers/create_join_relationship.ex
@@ -9,7 +9,7 @@ defmodule Ash.Resource.Transformers.CreateJoinRelationship do
 
   alias Ash.Dsl.Transformer
 
-  @extension Module.concat(["Ash", Resource, Dsl])
+  @extension Ash.Resource.Dsl
 
   def transform(_resource, dsl_state) do
     dsl_state