+++ CHANGELOG.md
@@ -5,6 +5,75 @@ See [Conventional Commits](Https://conventionalcommits.org) for commit guideline
 
 <!-- changelog -->
 
+## [v1.48.0-rc.6](https://github.com/ash-project/ash/compare/v1.48.0-rc.5...v1.48.0-rc.6) (2021-09-20)
+
+
+
+
+### Bug Fixes:
+
+* support `on` for global changes
+
+* return proper result when input is struct
+
+* remove File.read! from docs
+
+### Improvements:
+
+* skip resource action if no changes have been made
+
+* add `changing_attributes?/1` to determine if any attributes are chaning
+
+* add global changes
+
+## [v1.48.0-rc.5](https://github.com/ash-project/ash/compare/v1.48.0-rc.4...v1.48.0-rc.5) (2021-09-17)
+
+
+
+
+### Improvements:
+
+* upgrade docs/tooling for elixir_sense
+
+* set docs statically
+
+## [v1.48.0-rc.4](https://github.com/ash-project/ash/compare/v1.48.0-rc.3...v1.48.0-rc.4) (2021-09-17)
+
+
+
+
+### Improvements:
+
+* improvements for elixirsense integration
+
+## [v1.48.0-rc.3](https://github.com/ash-project/ash/compare/v1.48.0-rc.2...v1.48.0-rc.3) (2021-09-16)
+
+
+
+
+### Improvements:
+
+* transformer/extension improvements
+
+* add path to errors
+
+## [v1.48.0-rc.2](https://github.com/ash-project/ash/compare/v1.48.0-rc.1...v1.48.0-rc.2) (2021-09-15)
+
+
+
+
+### Bug Fixes:
+
+* set tenant properly on create interface
+
+* update type spec for Ash.Sort to include single atom instead of only list. (#263)
+
+### Improvements:
+
+* remove IO.inspect :*( and fix sobelow
+
+* support non-endpoint pubsub adapters
+
 ## [v1.48.0-rc.1](https://github.com/ash-project/ash/compare/v1.48.0-rc.0...v1.48.0-rc.1) (2021-09-13)